data/pyproject.toml
[tool.poetry]
name = "data"
version = "0.1.0"
description = ""
authors = ["Simon Kirby <s.kirby@exeter.ac.uk>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
numpy = "^1.26.4"
geojson = "^3.1.0"
psycopg2 = "^2.9.9"
netcdf4 = "^1.6.5"
pyyaml = "^6.0.1"
shapely = "^2.0.3"
pyproj = "^3.6.1"
rasterio = "^1.3.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
